saas-backend/requirements.txt
# Core FastAPI and ASGI dependencies
fastapi==0.115.12
uvicorn[standard]==0.34.2
starlette==0.46.2

# Database dependencies
databases[postgresql]==0.9.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.40
greenlet==3.2.1

# Pydantic and validation
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic-extra-types==2.10.3
pydantic_core==2.33.1
email-validator==2.2.0

# Security and authentication
bcrypt==4.3.0
cryptography==44.0.2
PyNaCl==1.5.0
python-multipart==0.0.20

# HTTP and networking
httpx==0.28.1
httpcore==1.0.8
websockets==15.0.1

# Utilities
python-dotenv==1.1.0
orjson==3.10.16
ujson==5.10.0
PyYAML==6.0.2

# Development and CLI tools (optional for production)
# fastapi-cli==0.0.7
# typer==0.15.2
# rich==14.0.0
# rich-toolkit==0.14.1

# Additional dependencies
anyio==4.9.0
sniffio==1.3.1
click==8.1.8
Jinja2==3.1.6
MarkupSafe==3.0.2